requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-tailwind==3.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
reportlab==4.4.5
sqlparse==0.5.3
tzdata==2025.2
